+++ include/buzzer.h
@@ -4,4 +4,5 @@ namespace Buzzer
 {
     void begin();
     void beep(unsigned long durationMs);
+    void doubleBeep(unsigned long durationMs, unsigned long gapMs);
 }